Gutted the database connection and API controller. Why: it wasn't being used, and it was too premature to implement it. It will not be hard to get it back in, but for the main focus is to get it working as is. Also refactored the codebase since a lot of folders were being removed. It is a lot cleaner now. All still works as expected, tests pass with flying colors.
