+++ src/main/java/com/loot/server/ServerApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@ComponentScan(basePackages = {"com.loot.server.api", "com.loot.server.socket", "com.loot.server.service", "com.loot.server.repositories"})
+@ComponentScan(basePackages = {"com.loot.server.service", "com.loot.server"})
 @SpringBootApplication
 @EnableScheduling
 public class ServerApplication {
