+++ src/main/java/com/loot/server/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.loot.server.socket;
+package com.loot.server;
 
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Configuration;
