--- src/main/java/com/loot/server/domain/entity/PlayerEntity.java
@@ -1,25 +1,0 @@
-package com.loot.server.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-public class PlayerEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "username")
-    private String name;
-
-    @Column(name = "image")
-    private String image;
-}
