+++ src/main/java/com/loot/server/domain/request/GamePlayer.java
@@ -3,7 +3,6 @@ package com.loot.server.domain.request;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.loot.server.domain.dto.PlayerDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
