+++ src/main/java/com/loot/server/domain/request/LobbyRequest.java
@@ -1,9 +1,7 @@
 package com.loot.server.domain.request;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.loot.server.domain.dto.PlayerDto;
 import lombok.*;
 
 @Data
