+++ src/main/java/com/loot/server/domain/request/PlayCardRequest.java
@@ -2,7 +2,6 @@ package com.loot.server.domain.request;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.loot.server.domain.dto.PlayerDto;
 import com.loot.server.domain.cards.PlayedCard;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
