--- src/main/java/com/loot/server/mapper/Mapper.java
@@ -1,8 +1,0 @@
-package com.loot.server.mapper;
-
-public interface Mapper<A, B> {
-
-    B mapTo(A a);
-
-    A mapFrom(B b);
-}
