--- src/main/java/com/loot/server/repositories/PlayerRepository.java
@@ -1,7 +1,0 @@
-package com.loot.server.repositories;
-
-import com.loot.server.domain.entity.PlayerEntity;
-import org.springframework.data.repository.CrudRepository;
-
-public interface PlayerRepository extends CrudRepository<PlayerEntity, Long> {
-}
