--- src/main/java/com/loot/server/socket/logic/IPlayerHandler.java
@@ -1,23 +1,0 @@
-package com.loot.server.socket.logic;
-
-import com.loot.server.domain.dto.PlayerDto;
-import com.loot.server.domain.cards.Card;
-
-public interface IPlayerHandler {
-
-    void addPlayer(PlayerDto player);
-
-    Boolean hasNextPlayer();
-
-    PlayerDto getNextPlayer();
-
-    Boolean addWinToPlayer(PlayerDto player);
-
-    void addPlayedCard(PlayerDto player, Card card);
-
-    Boolean readyUp(PlayerDto player);
-
-    void removePlayerFromRound(PlayerDto player);
-
-    void startNewRound();
-}
