+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/db
-spring.datasource.username=root
-spring.datasource.password=password
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.jpa.hibernate.ddl-auto=update
+#spring.datasource.url=jdbc:mysql://localhost:3306/db
+#spring.datasource.username=root
+#spring.datasource.password=password
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
