+++ src/test/java/com/loot/server/GameSessionTests/TestsForCardStack.java
@@ -1,7 +1,7 @@
 package com.loot.server.GameSessionTests;
 
 import com.loot.server.domain.cards.Card;
-import com.loot.server.socket.logic.impl.CardStack;
+import com.loot.server.logic.impl.CardStack;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
